fix(arc-400): pr remarks
