+++ src/modules/cp/components/ApproveRequestBlade/ApproveRequestBlade.const.ts
@@ -1,8 +1,5 @@
-import { format } from 'date-fns';
 import { i18n } from 'next-i18next';
 import { date, object, SchemaOf, string } from 'yup';
-
-import { getLocaleFromi18nLanguage } from '@shared/utils';
 
 import { ApproveRequestFormState } from './ApproveRequestBlade.types';
 
@@ -12,14 +9,4 @@ export const APPROVE_REQUEST_FORM_SCHEMA = (): SchemaOf<ApproveRequestFormState>
 		accessTo: date().required(i18n?.t('Een einddatum is verplicht.')),
 		accessRemark: string().optional(),
 	});
-};
-
-export const ApproveRequestAccessDateFormatter = (date?: Date): string => {
-	const locale = getLocaleFromi18nLanguage(i18n?.language || '');
-	return date ? format(date, 'P', { locale }) : '';
-};
-
-export const ApproveRequestAccessTimeFormatter = (date?: Date): string => {
-	const locale = getLocaleFromi18nLanguage(i18n?.language || '');
-	return date ? format(date, 'p', { locale }) : '';
 };
