+++ src/modules/cp/components/ApproveRequestBlade/ApproveRequestBlade.tsx
@@ -27,13 +27,13 @@ import { OPTIONAL_LABEL } from '@shared/const';
 
 import parentStyles from '../ProcessRequestBlade/ProcessRequestBlade.module.scss';
 
-import {
-	APPROVE_REQUEST_FORM_SCHEMA,
-	ApproveRequestAccessDateFormatter,
-	ApproveRequestAccessTimeFormatter,
-} from './ApproveRequestBlade.const';
+import { APPROVE_REQUEST_FORM_SCHEMA } from './ApproveRequestBlade.const';
 import styles from './ApproveRequestBlade.module.scss';
 import { ApproveRequestBladeProps, ApproveRequestFormState } from './ApproveRequestBlade.types';
+import {
+	formatApproveRequestAccessDate,
+	formatApproveRequestAccessTime,
+} from './ApproveRequestBlade.utils';
 
 const roundToNearestQuarter = (date: Date) => roundToNearestMinutes(date, { nearestTo: 15 });
 const defaultAccessFrom = (start: Date) => roundToNearestQuarter(start);
@@ -142,7 +142,7 @@ const ApproveRequestBlade: FC<ApproveRequestBladeProps> = (props) => {
 										maxDate={null}
 										{...field}
 										onChange={(date) => onFromDateChange(date, field)}
-										value={ApproveRequestAccessDateFormatter(field.value)}
+										value={formatApproveRequestAccessDate(field.value)}
 										selected={field.value}
 										customInput={
 											<TextInput iconStart={<Icon name="calendar" />} />
@@ -155,7 +155,7 @@ const ApproveRequestBlade: FC<ApproveRequestBladeProps> = (props) => {
 										minTime={defaultAccessFrom(now)}
 										{...field}
 										onChange={(date) => onFromDateChange(date, field)}
-										value={ApproveRequestAccessTimeFormatter(field.value)}
+										value={formatApproveRequestAccessTime(field.value)}
 										selected={field.value}
 										customInput={
 											<TextInput iconStart={<Icon name="clock" />} />
@@ -187,7 +187,7 @@ const ApproveRequestBlade: FC<ApproveRequestBladeProps> = (props) => {
 										minDate={accessFrom}
 										{...field}
 										onChange={(date) => onSimpleDateChange(date, field)}
-										value={ApproveRequestAccessDateFormatter(field.value)}
+										value={formatApproveRequestAccessDate(field.value)}
 										selected={field.value}
 										customInput={
 											<TextInput iconStart={<Icon name="calendar" />} />
@@ -204,7 +204,7 @@ const ApproveRequestBlade: FC<ApproveRequestBladeProps> = (props) => {
 										}
 										{...field}
 										onChange={(date) => onSimpleDateChange(date, field)}
-										value={ApproveRequestAccessTimeFormatter(field.value)}
+										value={formatApproveRequestAccessTime(field.value)}
 										selected={field.value}
 										customInput={
 											<TextInput iconStart={<Icon name="clock" />} />
