+++ src/modules/cp/components/ApproveRequestBlade/index.ts
@@ -1,1 +1,4 @@
 export { default as ApproveRequestBlade } from './ApproveRequestBlade';
+export * from './ApproveRequestBlade.const';
+export * from './ApproveRequestBlade.types';
+export * from './ApproveRequestBlade.utils';
