+++ src/modules/cp/const/requests.const.ts
@@ -1,17 +1,9 @@
 import { TabProps } from '@meemoo/react-components';
-import { differenceInDays, format, formatDistanceToNow } from 'date-fns';
 import { i18n } from 'next-i18next';
 import { NumberParam, StringParam, withDefault } from 'use-query-params';
 
+import { RequestStatus } from '@cp/types';
 import { SortDirectionParam } from '@shared/helpers';
-import { getLocaleFromi18nLanguage } from '@shared/utils';
-
-export const enum RequestStatus {
-	all = 'all',
-	open = 'open',
-	approved = 'approved',
-	denied = 'denied',
-}
 
 export const RequestTablePageSize = 20;
 
@@ -42,19 +34,4 @@ export const requestStatusFilters = (): TabProps[] => {
 			label: i18n?.t('modules/cp/const/requests___geweigerd'),
 		},
 	];
-};
-
-export const requestCreatedAtFormatter = (date: Date): string => {
-	const locale = getLocaleFromi18nLanguage(i18n?.language || '');
-
-	if (differenceInDays(new Date(), date) <= 5) {
-		return formatDistanceToNow(date, {
-			addSuffix: true,
-			locale,
-		});
-	}
-
-	return format(date, 'PPpp', {
-		locale,
-	});
 };
