+++ src/modules/shared/utils/i18n.ts
@@ -1,0 +1,10 @@
+import { Locale } from 'date-fns';
+import { nlBE } from 'date-fns/locale';
+
+export function getLocaleFromi18nLanguage(language: string): Locale {
+	switch (language) {
+		case 'nl':
+		default:
+			return nlBE;
+	}
+}
